fix: Don't name variables in ways that collide with DOM APIs
Assuming the FileReader API wasn't available, we were trying to create
a variable named FileReader. Due to hoisting, this would always
overwrite the original FileReader symbol when in a DOM context,
meaning this never worked. Now we only call FileAPI.FileReader via a
module resolution, which makes this work on both node and web.

Fixes #10